0.0.2.4

+++ KoalaDream/Assets/Scripts/Bank/BankDisplayView.cs
@@ -36,7 +36,7 @@ public class BankDisplayView : MonoBehaviour
         tweenScale = moneyDisplay.DOScale(new Vector3(1.1f, 1.1f, 1.1f), 0.1f).OnComplete(() => tweenScale = moneyDisplay.DOScale(defaultMoneyTableScale, 0.2f));
     }
 
-    public void SendMoneyDisplay(float money)
+    public void SendMoneyDisplay(int money)
     {
         textMoney.text = money.ToString();
     }
